backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
requests
httpx
web3
redis
python-multipart
websockets
py-clob-client
openai
